First pass to split code into different files
