+++ CMakeLists.txt
@@ -14,5 +14,10 @@ endif()
 
 set(LINK_LIBS ${LINK_LIBS} cfitsio ${HDF5_LIBRARIES})
 
-add_executable(hdf_convert main.cc)
+set(SOURCE_FILES
+    ${SOURCE_FILES}
+    main.cc
+    Image.cc)
+
+add_executable(hdf_convert ${SOURCE_FILES})
 target_link_libraries(hdf_convert ${LINK_LIBS})
